Fix `test_draw_gui`

+++ hotham/src/rendering/primitive.rs
@@ -109,7 +109,7 @@ impl Primitive {
 
         if let Some(iter) = reader.read_weights(0) {
             for t in iter.into_f32() {
-                joint_weights.push([t[0] as f32, t[1] as f32, t[2] as f32, t[3] as f32].into());
+                joint_weights.push([t[0], t[1], t[2], t[3]].into());
             }
         } else {
             for _ in 0..positions.len() {
